1、bug fix 屏幕坐标用int 2、序列化反序列化关闭文件

+++ demo/VS/unrealsoc-pure/unrealsoc/HSoftRaster/HMath/HVector2.h
@@ -9,18 +9,18 @@ namespace HSoftRaster
         {
         }
 
-        HVector2(float x, float y)
+        HVector2(signed int x, signed int y)
             : X(x),
               Y(y)
         {
         }
 
-        double X;
-        double Y;
+        signed int X;
+        signed int Y;
 
         static HVector2 GetZero()
         {
-            return HVector2(0.0f, 0.0f);
+            return HVector2(0, 0);
         }
     };
 }
